+++ driver_test.go
@@ -1487,6 +1487,9 @@ func TestPodmanDriver_ReadOnlyFilesystem(t *testing.T) {
 
 // check userns mode configuration (single value)
 func TestPodmanDriver_UsernsMode(t *testing.T) {
+	// TODO: run test once CI can use rootless Podman.
+	t.Skip("Test suite requires rootful Podman")
+
 	taskCfg := newTaskConfig("", busyboxLongRunningCmd)
 	taskCfg.UserNS = "host"
 	inspectData := startDestroyInspect(t, taskCfg, "userns")
@@ -1496,6 +1499,9 @@ func TestPodmanDriver_UsernsMode(t *testing.T) {
 
 // check userns mode configuration (parsed value)
 func TestPodmanDriver_UsernsModeParsed(t *testing.T) {
+	// TODO: run test once CI can use rootless Podman.
+	t.Skip("Test suite requires rootful Podman")
+
 	taskCfg := newTaskConfig("", busyboxLongRunningCmd)
 	taskCfg.UserNS = "keep-id:uid=200,gid=210"
 	inspectData := startDestroyInspect(t, taskCfg, "userns")