+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 IMPROVEMENTS:
 
+* config: Allow setting `pids_limit` option. [[GH-203](https://github.com/hashicorp/nomad-driver-podman/pull/203)]
 * runtime: Set mount propagation from TaskConfig [[GH-204](https://github.com/hashicorp/nomad-driver-podman/pull/204)]
+
 
 ## 0.4.1 (November 15, 2022)